PS interpreter - Check an operand exists before using it
Bug #706982 " Round 2: Segfault READ in real_param (iutil.c:861)"

In this case it is the sampled function code at fault, not the location
of the seg fault where the error occurs.

The sampled function code assumes there will be an operand on the stack
but in this case the function calls the 'clear' operator, so there is
nothing on the stack, which leads to the crash.

If this occurs, throw a stackunderflow error.