Bug 706979: Another stack block edge case in exec stack cleanup
In this case, the exec stack cleanup opproc house keeping entries are at the
bottom of one stack block, and the proc itself at the top of the previous
block. So by the time we call the cleanup proc, the block with its data has
been popped and freed.

So, we now pop to the bottom of the housekeeping entries, call the proc then
pop the proc entry.