PS interpreter - review use of check_estack()
The check_estack macro can call ref_stack_extend() in order to increase
the size of the exec stack. If this occurs any saved pointers to the
old exec stack will be invalidated.

Reviewing all the places where the macro gets used there are a few that
have pointers that may become stale after using the macro. This commit
refreshes them to avoid any errors.