FAPI: Fix type mismatch - int vs long
Passing a 64 bit long via pointer to a function that expects an int pointer
ends up causing problems on big endian platforms, as the "wrong" 32 bits of the
value get written to.

Use the correct variable types for the two specific cases.

Note: this should all be revised to use int32_t and co.... when time allows.