PS interpreter review colour code for stack pointers
No bug report but after the recent slew of fuzzing bugs exploiting
incorrect or missing stack checking, a desk check of the colour and
colour space code.

Mostly the code seems robust, there are a few improvements here; in a
couple of places add checks that the operand stack has at least the
expected number of operands. A number of places add checks that the
exec stack can hold an additional item before incrementing esp. Quite a
few places replace explicit decrement of esp with ref_stack_pop() in
order to ensure we don't fall off the bottom of a stack block.