PDF interpreter - fix image Lab space
Bug #706697 "JPX Lab spaces throw error and render blank"

Chris did the first part of the work here; there were a couple of typos
in pdfi_create_JPX_Lab() passing a pointer to a pointer to a PDF array
instead of simply a pointer to the array.

The second part of the fix is simply that the Range array for an Lab
space is optional and we were insisting one be present. If there is no
Range array, simply use the default.