Coverity ID #398206
I actually do not believe this is a real problem. Although we do
potentially copy uninitialised data to the ref's 'value', the type of
the ref is still set to t_null, so nothing should ever be consulting
the value anyway, as null objects don't have values.

But we can cheaply silence the error just by setting the value to 0.