Bug 707001: detect stack block underflow in sampled_data_continue()
sampled_data_continue() had code to detect the operand stack ending up in an
unexpected state after the sampling function being called. But that relies on
the overall stack depth, not the entries available on the current stack block.

So, add detection of underflow on the current op stack block, so we can return
the error, and have the calling code pop the stack block, and re-try.