PS interpreter - copy source data string for image operator (again)
Bug #706867 "heap use after free READ sample_unpack_8 / gx_image1_plane_data"

This revisits commit 09201ae46bcad1c6108956cb9a19f88f58e6c567 which
works fine but causes memory use to explode when used with type 3x
(soft mask) images with a file data source, and the pdfwrite family
of devices.

The way that the graphics library works with the PostScript interpreter
is more complex than I had realised, on debugging the code again it also
became clear that for PostScript we are completely reliant on the
garbage collector to clean up the string data passed into the graphics
library. I believe this is the source of the memory explosion when we
copy them into stable memory.

So in this case I've reverted the original commit and added a new flag
to gs_imaeg_next_planes() which explicitly transfers control of the
string lifetime to that routine. When this flag is true the string should
have been created in stable memory.

In this case we store an extra copy of the original data pointer and
size (because the 'source' string gets altered) so that we can properly
free the strings when we have control. The fact that the 'orig' pointer
is non-NULL and the size is not 0 are flags that we control these
strings and we must free them when we exhaust the data.

The fact that we've copied the data to stable memory means we no longer
have a problem when the restore discards the original strings, the fact
that we free the copied strings means we don't have to rely on the
garbage collector noticing they are no longer used and freeing them, and
the file data sources now work correctly without memory explosion.