Bug 707000: Protect against spurious grestores between image rows
In this case, a Postscript image defined with a procedure data source, and the
date source does a grestore (no matching gsave). This can cause vital things
like device halftones, color spaces, etc, to disappear with the reversion to a
previous graphics state.

To guard against that, record the gstate "level" when we begin the image, and
each time we pass image data to the graphics library, check that the current
gstate level is no lower than the initial one.

If it is, throw an error.