PS interrpreter - Make sure to use the system definition of findresource
Bug #707003 "Heap use after free WRITE of size 8 in indexed_cont (zcolor.c:4808)"

The problem arises because wee are trying to use a Pattern colour space
as the base space for an Indexed colour space, Pattern colour spaces
have a number of components defined as -1, which doesn't work well
when we try to push that many components onto the operand stack.

Pattern isn't a valid base space for Indexed colour spaces, and we do
check for that. The reason we can get here is because the specimen file:
1) redefines findresource as /Pattern
2) sets UseCIEColor to true

Because UseCIEColor is true the PostScript interpreter runs a small
chunk of PostScript code to turn the DeviceRGB alternate into a CIE
colour space. That code executes findresource, which pushes /Pattern
onto the operand stack. The result is we try to use /Pattern as the
base space for a /Indexed space.

Fixed simply by pulling findresource from systemdict, which can't be
redefined. While we're here, the code also uses 'stopped', so make sure
we use the system definition of that too.

Repeat for base spaces of DeviceGray and DeviceRGB.