PS interpreter - check there are operands before discarding them
Bug #707004 "Heap buffer overflow write of size 4 in ciebasecolor (zcolor.c:3513)"

This is another abuse of SetCIEColor. In this case the alternate for a
Separation space is defined as DeviceCMYK but because we have
UseCIEColor set to true that turns into a CIEBasedDEFG space. The tint
transform also executes a 'clear'.

Executing currentgray then handles the DEFG space and pops off 4 operands
from the operand stack, but doesn't check there are actually 4 present
before doing so, the 'clear' means there are not and so we fall off
the bottom of the operand stack.

The reason we discard the values is because the PLRM says that we should
return 0 for every component in this case, so we don't care what the
values are.

Fix by just checking to make sure there are 'n' operands (1, 3 or 4
depending on the CIE space) before discarding them.