graphics library - ensure pattern tile variables used as divisors are not zero
Bug #707005 "FPE clist_new_tile_params / clist_change_tile"

This is actually a reprise of an earlier report, #706969, with a patch
which I accepted at the time. Turns out this was insufficient.

This bug also has a patch but it is quite extensive and would
potentially have a negative impact on performance (many checks against
zero). Instead, when the tile is first initialised check the values
and if they are 0 use the NULL tile (because the tile won't actually
draw anything anyway).