PS interpreter - actually check the result of pow() for the exp operator
Bug #707014 "Segfault READ in is_tensor_done (gsfunc0.c:633)"

As usual, the actual cause is a long way from the error.

The 'exp' operator uses the C pow() function, which can return #INF if
the arguments are large. The check for this (where possible) was using
isinf(), but instead of checking the result it was checking the second
of the two operands on the stack....

Passing a #INF to the interpolation function causes it to try and access
invalid memory.