PS interpreter - don't copy gstate to itself
Bug #707016 "Segfault READ in cs_adjust_color_count (gxgstate.h:486)"

The problem occurs in gstate_copy() which has this comment:

    /*
     * It's OK to decrement the counts before incrementing them,
     * because anything that is going to survive has a count of
     * at least 2 (pto and somewhere else) initially.
     * Handle references from contents.
     */

Sadly that isn't true if pto and pfrom are the *same* graphics state!

We can get here with the gstate pointing to a colour space with a
reference count of 1 (just pointed to by the gstate itself) and in that
case counting it down first frees it...

Rather than mess with the reference counting, we can spot the fact that
we're copying a gstate to itself at a higher level and since this is a
no-op, just avoid the copy entirely.