PDF interpreter - apply text anti-aliasing to Tr 1, 2, 5 and 6
No bug report. This came up on Stack Overflow as a question and the
expectation doesn't seem unreasonable.

The basic problem is that text which is stroked, or filled and then
stroked, doesn't appear to get anti-aliased. This is because the
TextAlphaBits parameter only affects the 'fill' of text. In the case of
fill and then stroke, the fill is anti-aliased, but it's then over
written by the stroke.

This commit sets the GraphicsAlphaBits value to the TextAlphaBits value
for the duration of the text rendering, if they are not the same.

Note! This means that text which was formerly anti-aliased, because it
has a stroke, and GraphicsAlphaBits was set, now won't be, unless
TextAlphaBits is set.