Bug 707007/707015/707025: Don't leave a dangling pointer on the stack
At the end of a show operation, we free the enumerator associated with it, and
decrement the exec stack pointer - effectively popping the enumerator and
various bits of house keeping off the stack.

The problem is that in various places the garbage collector works entirely on
memory "clumps", and has no concept of where the top of the stacks actually are.

As a consequence the freed enumerator's memory can end up being enumerated
(marked) and relocated, which obviously leads to problems.

In the case of the test file for 707015 on my specific setup, for example, the
allocator's memory has be reused to hold a memory rendering device, the mark
and reloc corrupt that, which causes a crash when we come to free the device.
Note that exactly what gets corrupted depends very much on the memory use at
the time which is test file, configuration, platform etc specific.