Bug 707024: Don't pop stack entries for error returns
Returning an error to the interpreter loop should not pop items off the
stack(s), because we access the stack entries for error handling.

But, more importantly, returning an error code to the intepreter loop doesn't
trigger it to sync it's own stack references (because it expects the stacks to
remain unchanged, for error reporting). If, as in this case, popping the items
also causes a stack block to be popped, we can end up accessing freed memory.