Graphics library - validate the saved halftone before use.
Bug #707026 "Segfault READ in gx_dc_ht_binary_load (gxht.c:226)"

When we load a binary halftone, we use a stored pointer to retrieve the
'order'. This should always be one of teh halftones stored in the
graphics_state 'dev_ht' array, depending on the object type, which at
this point we don't know.

Because this is set earlier based on the object type, it should not be
possible to get here with an invalid halftone, but.... The image in this
case has a procedure data source, and that data source executes grestore
more times than gsave, until we fall off the top of the gstate stack.
The PLRM states that in this case we should copy the gstate from the
previous save state.

And when we do that, the halftone we are in the middle of using
disappears.

This commit checks the stored pointer against the dev_ht members to see
if the halftone is still valid (still in the gstate). If it isn't then
something went badly wrong, so generate an error and exit.

I have not altered gx_dc_ht_binary_load_cache() because firstly it
does not get called with a graphics state and secondly it is only
called when rendering so we sholdn't run into this problem (I hope).