DeviceN parameter handling - check number of colorants
Bug #707037 "Buffer overflow in DeviceN SeparationColorNames parameter"

The problem is that the DeviceN parameter handling has no protection
against adding SeparationColorNames multiple times. Because the
intention is that this should be executed from the command line it seems
that the original author simply assumed nobody would be silly enough
to execute it twice.

However it is still standard PostScript, and so can be executed from any
input file, so we need to guard against malicious execution.

This commit checks the number of existing inks (if any) and the number
of new inks and if the total exceeds the current maximum number of inks
supported by the device throws an error.

While making the change I spotted an incorrect parameter name in an
error message and fixed that too (clearly a copy/paste error).