Bug 707030: change how optimised stream access is initialised
The "optimised" stream access used by the token scanner uses local variables
that to access to the stream data.

There is a simple validity check that the current position pointer is less than
the end-of-data address, but when a stream is marked as invalid the current
data pointer in the stream is set to NULL, but the end pointer is untouched.
Thus the current position is always less than the end.

Change the local pointer initialisation so that if the current pointer is NULL,
so is the end pointer.