Bug 707027 - Don't leave a dangling opinter on the exec stack
As per commit e7b9aae979e45843a96b862439827682bcfb6ad1 we explicitly
free an enumerator structure, but because GC can end up processing an
entire clump, this can mean that the ref to the struct is processed
even though it is no longer on the stack as such.

Make the ref a NULL to prevent problems.