PostScript interpreter - Null dangling references on stack
In the wake of bug #707007 it is clear that due ot the way GC works on
stack clumps of memory, we cannot leave references to structures which
have been explicitly freed on the stacks.

So far this only seems to be a problem with the exec stack, because
objects on the operand stack are left to GC to free.

This commit is the result of reviewing all the places I could find where
we allocate a structure on the stack, and then free it explicitly in C
rather than leaving it to the garbage collector.