mswinpr2 device - don't use unprotected unlink
Bug #707049 " "mswinpr2" device can delete arbitrary files"

The use of 'unlink' instead of the gp_ layer means that a malicious
PostScript program could delete files (at the user level of the
Ghostscript process).

This commit uses gp_unlink instead to prevent access to unauthorised
files.

In practice I believe this code is pointless, gdev_prn_open() does not
create a temporary file for me. I intend to remove the code in a
following commit, this commit is to enable us to pull a lower risk
change into the release process currently underway.