mswinpr2 device - remove pointless code
As I mentioned in the previous commit, for bug 707049, the code using
the unlink function appears to be pointless. gdev_prn_open() does not,
as far as I can tell, open a temporary file which we need to later
delete.

So remove the code; we won't put this in the current release (which
already has the mitigation using gp_unlink) in case a problem shows up
in the meantime.