PS interpreter - fix copy with a count of 0
Bug #707050 " "0 copy" terminates with stackunderflow if operand stack is empty"

This was part of a review of operators and operand checking, commit
5a2955e26488afc71369b93c189ce0c5eef7eeec

The copy operator wasn't previously checking that it had at least one
operand before checking the operand type, but the fix assumed that copy
would always get at least two operands, which is not the case if there
is a single numeric operand representing the number of objects to copy
and the operand value is 0.

So return the check for two operands to the position it was before, and
add a new check to see that we have at least one operand before testing
the type of that operand.