OSS-Fuzz #60470
The problem is due to the fact that the original PDF has a recursive
type 3 font, and a pattern. Although we don't fall into the recursion
trap, we do end up executing several levels of recursion.

We scale the CTM by a factor of 100 when capturing type 3 CharProc
routines in pdfwrite, I'm not certain why. If fonts are recursive then
each time we start capturing a CharProc we scale the CTM by a further
factor of 100.

We also multiply the device width and height by 100 at the same time. In
the case of this test file that causes the height to overflow an integer
and become negative. Later we try to fill with a pattern and the pattern
clamping code tries to apply a negative height, which makes no sense
at all.

There's really no point in scaling the CTM more than once when capturing
CharProcs, so in this commit I limit the scaling so that it is only
performed if we have not already scaled the CTM. We only undo the
scaling when we exit the original CharProc.

This avoids the overflow, but I added a check for that too, just in
case.