pdfwrite - Improve handling of mesh shadings
Bug #706852 "shfill not shown in pdf output"

The source of the problem is the limited range of real numbers in old
versions of Acrobat. The input file has a type 5 (lattice form Gouraud
shaded triangle) with an array data source for the vertex data.

Because of the scaling involved, this results in co-ordinates for the
vertices which are comparatively large (> 100,000).

Early versions of Acrobat (NOT PDF) had an implementation limit of
32,767 for real numbers, noted in the implementation limits of the
PDF 1.3 specification. In fact Acrobat 4 actually has an implementation
limit of 16,384 on real numbers.

This does, of course, rather make a mockery of the 24 and 32-bit per
co-ordinate cases....

In order to work with this the code in pdfwrite was clamping the
co-ordinates to 16,384, which means that the entire shading was being
rendered to a degenerate point.

No, in recent versions of Acrobat (PDF 1.5, so Acrobat 6 or better) this
implementation limit has been raised to something more reasonable
(1x10^38) so it would seem that we could essentially remove this limit.

Unfortunately when the PDF/A specification was written the governing
group chose to enshrine the Acrobat limitation in the specification and
so PDF/A-1 has a limit on real numbers of 32,767. As above, later
versions of PDF/A raise this.

This commit checks to see the compatibility level that pdfwrite is
conforming to; if it is less than 1.5 (selecting -dPDFA sets the
compatibility level to 1.3 behind the scenes) and we have an array
data source, then we check all the co-ordinates to see if any will
exceed the maximum value. If so then we choose to rescale the
co-ordinates. We do that by modifying the /Matrix applied to the Pattern.

Because the values are absolute co-ordinates, this doesn't affect the
general rendering of the pattern. We do lose accuracy by doing this
but it is the only alternative which doesn't lead to us losing the
shading either in part or in whole.

We don't have many test files for this (mesh shading with array data
source) so I haven't been able to test this as thoroughly as I would
like but my hand crafted files work as does the original test file,
even when creating a PDF/A-1b file.

Somewhat surprisingly this does lead to some minor differences in
shadings in a few test file, because we are no longer clamping values.
None of these differences are visible to the eye.