Coverity ID's 398348 and 398349
Complaints that we are testing pgs against NULL after we have already
dereferenced it.

In one case the test is clearly testing the wrong variable and has
been removed as I don't know what it should be (if anything).

In the second case I've moved the dereference after the NULL test as I
suspect the NULL test may be important.

This should resolve both cases.