PS interpreter - limit recursion checking colour arrays
Bug #707081 "Stack overflow in array_get"

The file is, of course, crazy. It executes a setscreen, where the screen
procedure executes setcolorspace, with a /Separation space, and the
tint transform of the separation space stores an object in an array and
leaves that array on the stack.

The upshot of this is to create a massively nested array:
[[[[[[[[[[...[0]]]]].....]
which ends up being set as the 'PostScript' colour space. This doesn't
cause a problem.

However, the settransfer is run in a repeat loop. The second time we
call setscreen it starts all over again nesting the arrays still deeper.
Eventually we reach a point where the code which tests two colour spaces
to see if they are equal  recurses deeply enough to exhaust the C
exec stack.

Avoid this by simply limiting the depth of recursion. If we exceed the
recursion depth then we simply assume the spaces are not the same. This
should not cause any limitation with real files (because real colour
spaces don't nest arrays that deeply) and avoids the problem here.