Bug 707092: Limit the stack sanitizing depth
When we throw a stackoverflow error we store the contents of the op stack in
an array, so error handlers can access it.

That opened a security problem that procedures stored on the stack could be
accessed and operators pulled from them. So we "santize" them by replacing
real operators with strings representing their names.

But a circular reference can cause the recursion to overflow the C stack.

Limit the depth we recurse to, and after that, replace the offending array(s)
with a null object.