graphics library - fix ICC parameter parsing
Bug #707090 "Segfault READ in gx_default_put_icc_colorants (gsdparam.c:1546)"

The parsing of parameters here was wrong. If we got an error trying to
read one of the parameters (eg ICCOutputColors) then the return code
would not be 1 and the code proceeded as if the parameter was properly
parsed. Since it was not the code to set the parameter failed trying to
read from uninitialised memory.

There were a number of similar cases.