PS interpreter - pop exec stack on error sampling function
Bug #707091 "Segfault READ in gs_function_Sd_init (gsfunc0.c:1463)"

We should not get to the sampled function initialisation. We are in a
'nested' sampling; when the inner function throws an error it fails to
clean up the exec stack. When we close the outer function the enumerator
we think is on the exec stack is the wrong one and has already been
released.

Trying to use the freed enumerator causes a crash.