graphics library - don't flush a (writeable) stream if it is closed
Bug #707086 "Segfault READ in gp_fflush (gp.h:296)"

The test file establishes a data source, then duplicates the file
object, then establishes a filter on that file object. This leaves two
file objects on the stack, the data source and the filter.

Then the file exchanges the two objects and calls closefile, which
closes the data source. This violates the principle on page 124 of the
PLRM which states:

"Once a program has begun reading from or writing to a filter, it should not
attempt to access the underlying file in any way until the filter has been closed."

However, we should not crash, so this commit simply checks the status
of the source file when closing a filter and, if it is EOFC (closed for
a writeable file) then we don't try to flush it.