PDF interpreter - Don't try and use a closed input stream
Bug #707099 " Segmentation fault"

The file is corrupted and pdfi_set_input_stream() returns an error. This
causes the PostScript wrapper code to close the input file. But it was
not clearing the main stream in the PDF context.

Later, when closing the PDF, we would detect the error and try to
retrieve the Producer from the file, but because we'd closed the file
that caused a seg fault.

Fix two ways; in PostScript don't try and write errors if the input
stream of file is NULL. Also NULL the Ghostscript stream object in the
PDF context (this is safe, because we've just freed it). In PDF; don't
try and dereference an object if the main stream, or the underlying
Ghostscript stream object, is NULL.