Bug 707118: Maintain consistent global mode through screen sampling
Previously, we'd establish the global/local mode based on the global mode of
the spot procedures, and used that to allocate the halftone and device
halftone. But we didn't change the currentglobal. That caused problems with
garbage collection happening midway through the sampling.

Record currentglobal, set it so it is consistent through the sampling, and set
it back after sampling.