ps2write - free correct copy of PSPageOptions
Bug #707148 "Double free or corruption crash"

If we we were given a PSPageOptions parameter, but chose not to
replace the existing one because LockDistillerParams was true, then we
freed the wrong memory. We were freeing the stored options in the
device, instead of the newly supplied options.

While we're making changes, add some minor 'safety first' changes to
initialise the new params and ensure they start unallocated and only
free the new paramters if they differ from the stored ones.