Graphics library - Fix boundary check in function evaluation
Bug #707128 "Floating Point Exception in fn_PtCr_evaluate (gsfunc4.c:329)"

There is a check in the function evaluation to avoid an integer overflow
when dividing min_int by -1. The minimum 32-bit value is -2147483648
which, when divided by -1, becomes 2147483648 and max_int is
2,147,483,647, so we get a floating point exception.

Unfortunately, the check was performed *after* the division, by which
time it's too late!