PS interpreter - ensure object is a dictionary before use
Bug #707121 "Segfault READ in dict_next (idict.c:882)"

I am unable to reliably reproduce the problem in the bug report, some
builds reproduce it, some don't. Rebuild and the problem goes away. This
makes it all but impossible to work on these bugs.

In this case I had a fix that addresses the symptom rather than the
cause, and wanted to investigate the cause further. However I'm now
unable to reproduce the problem on any OS.

So I'm committing this fix anyway, which at least resolves the crash
even though I think there is an underlying problem.