PDF interpreter - Have PDFFitPage account for any existing /Rotate
Bug #707168 "Landscape pages are rotated by 180° with dPDFFitPage parameter"

The specimen file is actually a portrait page which has a /Rotate key
with a value of 90. We ignore the /Rotate key, as we want to fit the
page to the media whichever way leads to the least scaling.

However, this can lead to us apparently rotating the content by 180
degrees whether we rotate by 90 or 270, if the input file has a /Rotate
which happens to differ from what we use.

So if we are rotating the content to fit the requested media, look at
the Rotate key, if present, and use that as a hint whether we dhould
rotate clockwise or counter clockwise.