Bug 707122: Mark scanner_state ref with correct vm mode
When a stream buffer refill in the token scanner requires a call out to the
interpreter, as it does with a procedure data source, a scanner state object
is allocated in the current vm mode, and pushed onto the exec stack, along
with, amongst some other objects, the data source procedure.

If this process repeats, the same scanner state object will be reused until the
EOF condition or the scanning is otherwise completed.

A procedure data source for a file object (filter) can feasible change the
vm mode between local and global.

In which case, reusing the scanner state, pushing it onto the exec stack and
marking the stack ref with the currentvm attribute means the attribute and the
actual object do not match.

In event we are reusing an existing scanner state, make sure the ref is marked
as the same vm mode as the actual object it references.

This is applied to both regular scanning for the "token" scanning, and the
internally used "tokenexec" scanning, use specifically for executable file
objects.