PCL interpreter - Don't use the special handling for pattern transparency when the pattern is solid black
Bug 707105 "GhostPCL pdfwrite inverts color for raster image"

When encountering transparency while printing raster data, the PCL code
has to generate masks. This works when rendered as a bitmap but can lead
to problems in the pdfwrite device. As long as the used pattern does not
contain transparent pixels, there is no need to generate transparency
masks, regardless of the value of pattern_transparency. This way, the
result will be created correctly even with pdfwrite.