PDF interpreter - loosen text rendering warning
Bug #707281 "Over enthusiastic warning on text rendering modes"

The file sets the text rendering mode to a clipping mode, then exits the
text block, begins a new text block and sets the rendering mode to 0.

The text rendering mode can be set outside the text block, and really
should be, because switching from a clipping text rendering mode to a
non-clipping one is an error. However, because in this case no actual
text has been drawn yet it is functionally equivalent to changing the
mode before starting the text block, and so really shouldn't trigger
the warning.

Set a flag when we draw text in a clipping mode, and clear it when we
execute an 'ET'. Extend the warning test to additionally check if the
flag is true ie (text has been drawn in a clipping mode).