Coverity ID 403071
Potential use of uninitialised variable bubbleBuffer if an earlier
allocation failed.

At the same time; ensure the device stored pointers are NULL after we
free the data. Ensure BlockHeight cannot overflow. Also make sure
BlockWidth is 'reasonable'.

Although BloackWidth cannot overflow in exactly the same way we do use
it as a multiplier in many places which could themselves overflow and
are unchecked. Adding comprehensive checking would slow the performance,
and besides is more effort than we can justify for the contrib devices.

We don't own the copyright for these devices, so we need to keep
amendments as minimal as possible.