PS interpreter - remove some more functions
This is due to bug #707225 which used putdeviceprops to set a pdfmark
instead of going through the pdfmark function. This bypassed the
checking in that function and allowed an invalid parameter set that
would otherwise have been caught.

This commit removes putdeviceprops and getdeviceprops to avoid this
and additionally removes finddevice. There's no real call to use these
at all, since their action can be duplicated using setpagedevice and
currentpagedevice which are standard PostScript operators.

While we're in here making changes; remove copydevice and
findprotodevice (which is unused anyway)

This has meant modifying a couple of our lib files which were using
finddevice and putdeviceprops. I've tested these as well as I can given
that I don't have an Epson Stylus Colour printer. I did check the
eps2epsi.ps code with a non-EPS input file, and that worked correctly.

Also modify the documentation