Bug 707237/707244: Fix CMap parsing for begin/endbfrange
In the pdfi code, I had missed the incompatibility of the begin/endbfrange
handling of arrays between the "real" CMap spec, and what the original PDF
ToUnicode CMap implementation permitted.

In the CMap spec, bfranges mapping to arrays were restricted to glyph names
in the array - basically allowing for Type 1/2 fonts to be used as descendants
of a Type 0 (something not permitted in PDF).

The original ToUnicode implementation allowed glyph indices to be used in the
array, and I'd failed to take that into account.