PDF interpreter avoid loop with broken PDF file and PDFSTOPONERROR
Bug #707248 "The program seems to be stuck in a loop"

While repairing the file, and looking for an 'endobj', if we encounter
a delimiter (in this case ']') and PDFSTOPONERROR is true we return an
error and, crucially 'unread' the byte. The repair code, however,
ignores almost all errors because it is, after all, repairing a broken
file, so we expect errors.

But because it doesn't consume the byte we end up reading it again, and
again.

In this commit we consume a byte from the input when this occurs, and
if PDFSTOPONERROR is true, and the xref is broken, then we don't attempt
to repair the file anyway.