oss-fuzz 63013: pdfwrite clean up on type 3 image error
This highlighted two problems cleaning up after an error initialising to handle
a type 3 image.

First, only one of two interim devices was being freed, and as the remaining
device was a forwarding device that forwarded to the freed device, when the
garbage collector freed that remaining device, accessing the target device
caused use-after-free problems.

Secondly, the streams for the compression filter(s) were not being freed, and
left in an unknown state (by the error).

This fixes both those problems.