PDF interpreter - correct the usage of the loop detector
If we had a Font Resource dictioanry defined, but it was empty, then
we would not initialise the loop detector (because we didn't check any
fonts from the dictionary, obviously) but we would attempt to clear it
back down.

This leadds to a warning in a debug build. It is entirely harmless
currently but we shouldn't do that. This commit moves the cleanup so
that we only clean up the loop detector if we initialised it.