PDF interpreter - avoid potential seg fault
This exhibited with the file for OSS-fuzz #63013, probably because
of memory corruption. Fixing the code makes this problem go away, and
it wasn't the bug reported by OSS-fuzz, but it's worth the extra
check before dereferencing, just to avoid a seg fault.

It will only happen once at the end of the file.