PDF interpreter - handle illegal missing Count in Outlines
Bug #707228 "pdfwrite flattens outline hierarchy of some files"

The file is incorrectly constructed, it has an Outline entry which has
descendants, but is missing the (required in this case) /Count entry.

We were using the presence of /Count to indicate descendants, so the
missing entry caused us to 'flatten' the heirarchy.

We now use the presence of /First to indicate descendants. In addition,
assume that a missing or 0 Count indicates a closed node (ordinarily we
would use the Count being negative to indicate a closed node).