Bug 707205: Fix SEGV in pdf14_copy_planes.
We are arriving in pdf14_copy_planes without having a pdf14
context set up. As far as I can tell, the context is supposed
to be set up "just in time" when such things are called.

Most of the time we'll have gone through pdf14_fill_path or
something first, which will have done the required setup,
but in this particular case, the first pdf14 operation for
this band is copy_planes.

Accordingly, I've copied the init routine in here from
pdf14_fill_path, and this seems to solve the bug.