Graphics library - improved fix for OSS-fuzz #63185
We shouldn't decrement pcs unless we incrememnted it, the only way to
know if we incremented it is if penum->pcs points to it. So instead
of decrementing pcs, decrement penum->pcs.

Initialise penum->pcs to NULL because there are failure routes which
can get here without having set penum->pcs.