PDF interpreter - allow XRefStm in hybrid to override free entries
Bug #707260 "Hybrid file not working properly"

The file is a hybrid PDF file (PDF 1.4 & PDF 1.5) where some entries
are 'hidden'. They are defined in the PDF 1.5 XRefStm but not in the
PDF 1.4 standard xref.

This of course means that a PDF 1.4 consumer cannot access those objects
an to my mind makes the hybrid file pointless.

Anyway, the hidden entries are defin in the standard xref as free
entries, and we were not permitting the real entries in the XRefStm to
replace them, because the object entries were defined.

Simply check to see if the entry is free, and if it is permit the
XRefStm to override it.