memory device - prevent integer overflow in raster calculations
Bug #707266 "Image Device: Integer overflow in raster size calculation"

In the "gdev_mem_bits_size" function, the result of the "width * depth" expression is of type
int, and with large values it can easily overflow the range of an int. So, in the end, less
than the required amount of memory is allocated for the image data.

I added a typecast to size_t to avoid truncation of the result on systems where int is
smaller than size_t. And check of the values for multiplication and addition.

I also used dynamically calculated "alignment" variable when determining the limit.