Remove the 'makeimagedevice' operator under SAFER
Bug 707266 abused the operator, and I only gave it a reprieve on the
last round of non-standard operator removal because we document its
use.

We've now agreed to remove it when SAFER is true because it isn't really
used and it is always a potential danger.

I've updated the documentation to indicate it is only available when
NOSAFER is set, and updated the two utility programs (both for PCL)
which used it, to indicate that these programs must be run using NOSAFER