PDF interpreter - optionally honour PDF Permissions
The old PDF interpreter had commented-out code to honmour the 'print'
flag in encrypted PDF files. This was (presumably inadvertantly)
removed entirely in commit 49a4fede913a12a61fd6649898cf5999030f4b69
which meant it did not get carried over into the new interpreter.

This commit adds code which must be enabled at compile-time by adding
a compiler pre-processor flag USE_PDF_PERMISSIONS. If compiled in
the interpreter will refuse to process any file which does not have the
'print' flag set. In addition, if the print flag is set, but the
'faitfhul digital copy' is not set and the output device is a high level
device (eg pdfwrite) then we will refuse to process the file.